requirements.txt
python-dotenv==0.21.1
Pillow==10.0.0
loguru==0.6.0
pandas==2.0.3
openpyxl==3.1.2